+++ src/rdt/rdt.hpp
@@ -17,36 +17,12 @@
 
 #include <vector>
 
-#include "pqos.h"
+#include "pqos.hpp"
 
 #include "snap/plugin.h"
 #include "snap/config.h"
 #include "snap/metric.h"
 
-class PQOSInterface
-{
-public:
-    virtual ~PQOSInterface() {}
-    virtual int pqos_init(const pqos_config *config) = 0;
-    virtual int pqos_cap_get(const struct pqos_cap **cap, const struct pqos_cpuinfo **cpu) = 0;
-    virtual int pqos_fini() = 0;
-    virtual int pqos_mon_start(const unsigned num_cores, const unsigned *cores, const enum pqos_mon_event event, void *context, struct pqos_mon_data *group) = 0;
-    virtual int pqos_mon_reset() = 0;
-    virtual int pqos_mon_stop(struct pqos_mon_data *group) = 0;
-    virtual int pqos_mon_poll(struct pqos_mon_data **groups, const unsigned num_groups) = 0;
-};
-
-class PQOS : public PQOSInterface
-{
-public:
-    int pqos_init(const pqos_config *config);
-    int pqos_cap_get(const struct pqos_cap **cap, const struct pqos_cpuinfo **cpu);
-    int pqos_fini();
-    int pqos_mon_start(const unsigned num_cores, const unsigned *cores, const enum pqos_mon_event event, void *context, struct pqos_mon_data *group);
-    int pqos_mon_reset();
-    int pqos_mon_stop(struct pqos_mon_data *group);
-    int pqos_mon_poll(struct pqos_mon_data **groups, const unsigned num_groups);
-};
 
 namespace rdt
 {
